Renamed global logging hook to avoid confusion with new Winston logger file.
