+++ generators/app/templates/src/hooks/log.js
